chore: move queries back into server.go file
