+++ pkg/server/server.go
@@ -10,6 +10,8 @@ import (
 	grpc_ctxtags "github.com/grpc-ecosystem/go-grpc-middleware/tags"
 	"github.com/oklog/ulid/v2"
 	"github.com/openfga/openfga/internal/gateway"
+	"github.com/openfga/openfga/internal/graph"
+	"github.com/openfga/openfga/internal/validation"
 	"github.com/openfga/openfga/pkg/encoder"
 	"github.com/openfga/openfga/pkg/logger"
 	httpmiddleware "github.com/openfga/openfga/pkg/middleware/http"
@@ -81,6 +83,96 @@ func New(dependencies *Dependencies, config *Config) *Server {
 	}
 }
 
+func (s *Server) ListObjects(ctx context.Context, req *openfgapb.ListObjectsRequest) (*openfgapb.ListObjectsResponse, error) {
+
+	targetObjectType := req.GetType()
+
+	ctx, span := tracer.Start(ctx, "ListObjects", trace.WithAttributes(
+		attribute.String("object_type", targetObjectType),
+		attribute.String("relation", req.GetRelation()),
+		attribute.String("user", req.GetUser()),
+	))
+	defer span.End()
+
+	storeID := req.GetStoreId()
+	modelID := req.GetAuthorizationModelId()
+
+	typesys, err := s.typesystemResolver(ctx, storeID, modelID)
+	if err != nil {
+		if errors.Is(err, typesystem.ErrModelNotFound) {
+			return nil, serverErrors.AuthorizationModelNotFound(modelID)
+		}
+	}
+
+	if !typesystem.IsSchemaVersionSupported(typesys.GetSchemaVersion()) {
+		return nil, serverErrors.ValidationError(typesystem.ErrInvalidSchemaVersion)
+	}
+
+	q := &commands.ListObjectsQuery{
+		Datastore:             s.datastore,
+		Logger:                s.logger,
+		ListObjectsDeadline:   s.config.ListObjectsDeadline,
+		ListObjectsMaxResults: s.config.ListObjectsMaxResults,
+		ResolveNodeLimit:      s.config.ResolveNodeLimit,
+		TypesystemResolver:    s.typesystemResolver,
+	}
+
+	return q.Execute(ctx, &openfgapb.ListObjectsRequest{
+		StoreId:              storeID,
+		ContextualTuples:     req.GetContextualTuples(),
+		AuthorizationModelId: modelID,
+		Type:                 targetObjectType,
+		Relation:             req.Relation,
+		User:                 req.User,
+	})
+}
+
+func (s *Server) StreamedListObjects(req *openfgapb.StreamedListObjectsRequest, srv openfgapb.OpenFGAService_StreamedListObjectsServer) error {
+	ctx := srv.Context()
+	ctx, span := tracer.Start(ctx, "StreamedListObjects", trace.WithAttributes(
+		attribute.String("object_type", req.GetType()),
+		attribute.String("relation", req.GetRelation()),
+		attribute.String("user", req.GetUser()),
+	))
+	defer span.End()
+
+	storeID := req.GetStoreId()
+
+	modelID, err := s.resolveAuthorizationModelID(ctx, storeID, req.GetAuthorizationModelId())
+	if err != nil {
+		return err
+	}
+
+	model, err := s.datastore.ReadAuthorizationModel(ctx, storeID, modelID)
+	if err != nil {
+		if errors.Is(err, storage.ErrNotFound) {
+			return serverErrors.AuthorizationModelNotFound(req.GetAuthorizationModelId())
+		}
+
+		return serverErrors.HandleError("", err)
+	}
+
+	typesys := typesystem.New(model)
+
+	for _, ctxTuple := range req.GetContextualTuples().GetTupleKeys() {
+		if err := validation.ValidateTuple(typesys, ctxTuple); err != nil {
+			return serverErrors.HandleTupleValidateError(err)
+		}
+	}
+
+	q := &commands.ListObjectsQuery{
+		Datastore:             s.datastore,
+		Logger:                s.logger,
+		ListObjectsDeadline:   s.config.ListObjectsDeadline,
+		ListObjectsMaxResults: s.config.ListObjectsMaxResults,
+		ResolveNodeLimit:      s.config.ResolveNodeLimit,
+		TypesystemResolver:    s.typesystemResolver,
+	}
+
+	req.AuthorizationModelId = modelID
+	return q.ExecuteStreamed(ctx, req, srv)
+}
+
 func (s *Server) Read(ctx context.Context, req *openfgapb.ReadRequest) (*openfgapb.ReadResponse, error) {
 	tk := req.GetTupleKey()
 	ctx, span := tracer.Start(ctx, "Read", trace.WithAttributes(
@@ -116,6 +208,98 @@ func (s *Server) Write(ctx context.Context, req *openfgapb.WriteRequest) (*openf
 		AuthorizationModelId: modelID,
 		Writes:               req.GetWrites(),
 		Deletes:              req.GetDeletes(),
+	})
+}
+
+func (s *Server) Check(ctx context.Context, req *openfgapb.CheckRequest) (*openfgapb.CheckResponse, error) {
+	tk := req.GetTupleKey()
+	ctx, span := tracer.Start(ctx, "Check", trace.WithAttributes(
+		attribute.KeyValue{Key: "object", Value: attribute.StringValue(tk.GetObject())},
+		attribute.KeyValue{Key: "relation", Value: attribute.StringValue(tk.GetRelation())},
+		attribute.KeyValue{Key: "user", Value: attribute.StringValue(tk.GetUser())},
+	))
+	defer span.End()
+
+	if tk.GetUser() == "" || tk.GetRelation() == "" || tk.GetObject() == "" {
+		return nil, serverErrors.InvalidCheckInput
+	}
+
+	storeID := req.GetStoreId()
+	modelID := req.GetAuthorizationModelId()
+
+	typesys, err := s.typesystemResolver(ctx, storeID, modelID)
+	if err != nil {
+		if errors.Is(err, typesystem.ErrModelNotFound) {
+			return nil, serverErrors.AuthorizationModelNotFound(modelID)
+		}
+	}
+
+	if !typesystem.IsSchemaVersionSupported(typesys.GetSchemaVersion()) {
+		return nil, serverErrors.ValidationError(typesystem.ErrInvalidSchemaVersion)
+	}
+
+	if err := validation.ValidateUserObjectRelation(typesys, tk); err != nil {
+		return nil, serverErrors.ValidationError(err)
+	}
+
+	for _, ctxTuple := range req.GetContextualTuples().GetTupleKeys() {
+		if err := validation.ValidateTuple(typesys, ctxTuple); err != nil {
+			return nil, serverErrors.HandleTupleValidateError(err)
+		}
+	}
+
+	checkResolver := graph.NewLocalChecker(
+		storage.NewCombinedTupleReader(s.datastore, req.ContextualTuples.GetTupleKeys()),
+		s.typesystemResolver,
+		checkConcurrencyLimit,
+	)
+
+	resp, err := checkResolver.ResolveCheck(ctx, &graph.ResolveCheckRequest{
+		StoreID:              storeID,
+		AuthorizationModelID: modelID,
+		TupleKey:             req.GetTupleKey(),
+		ContextualTuples:     req.ContextualTuples.GetTupleKeys(),
+		ResolutionMetadata: &graph.ResolutionMetadata{
+			Depth: s.config.ResolveNodeLimit,
+		},
+	})
+	if err != nil {
+		if errors.Is(err, graph.ErrResolutionDepthExceeded) {
+			return nil, serverErrors.AuthorizationModelResolutionTooComplex
+		}
+
+		return nil, serverErrors.HandleError("", err)
+	}
+
+	res := &openfgapb.CheckResponse{
+		Allowed: resp.Allowed,
+	}
+
+	span.SetAttributes(attribute.KeyValue{Key: "allowed", Value: attribute.BoolValue(res.GetAllowed())})
+	return res, nil
+}
+
+func (s *Server) Expand(ctx context.Context, req *openfgapb.ExpandRequest) (*openfgapb.ExpandResponse, error) {
+	tk := req.GetTupleKey()
+	ctx, span := tracer.Start(ctx, "Expand", trace.WithAttributes(
+		attribute.KeyValue{Key: "object", Value: attribute.StringValue(tk.GetObject())},
+		attribute.KeyValue{Key: "relation", Value: attribute.StringValue(tk.GetRelation())},
+		attribute.KeyValue{Key: "user", Value: attribute.StringValue(tk.GetUser())},
+	))
+	defer span.End()
+
+	storeID := req.GetStoreId()
+
+	modelID, err := s.resolveAuthorizationModelID(ctx, storeID, req.GetAuthorizationModelId())
+	if err != nil {
+		return nil, err
+	}
+
+	q := commands.NewExpandQuery(s.datastore, s.logger, s.typesystemResolver)
+	return q.Execute(ctx, &openfgapb.ExpandRequest{
+		StoreId:              storeID,
+		AuthorizationModelId: modelID,
+		TupleKey:             tk,
 	})
 }
